requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
openai
jinja2
pyconnectwise